# synapse.ai---lead-radar-&-code-defense/backend/tests/snapshots/openapi.json
{
 "components": {
  "schemas": {
   "BenchmarkSuiteResponse": {
    "description": "Measured benchmark facts. Any value that was not measured is null.",
    "properties": {
     "createdAt": {
      "format": "date-time",
      "title": "Createdat",
      "type": "string"
     },
     "cyclomaticComplexity": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cyclomaticcomplexity"
     },
     "executionTimeMs": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Executiontimems"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "repoId": {
      "format": "uuid",
      "title": "Repoid",
      "type": "string"
     },
     "secVulnerabilitiesCount": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Secvulnerabilitiescount"
     },
     "suiteLogOutput": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Suitelogoutput"
     },
     "testCoveragePct": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Testcoveragepct"
     }
    },
    "required": [
     "id",
     "repoId",
     "createdAt"
    ],
    "title": "BenchmarkSuiteResponse",
    "type": "object"
   },
   "CandidateRepoDetailResponse": {
    "properties": {
     "benchmark": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/BenchmarkSuiteResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "branch": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Branch"
     },
     "candidateAvatarUrl": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Candidateavatarurl"
     },
     "candidateGithubHandle": {
      "title": "Candidategithubhandle",
      "type": "string"
     },
     "candidateName": {
      "title": "Candidatename",
      "type": "string"
     },
     "commitSha": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Commitsha"
     },
     "evaluation": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CodeEvaluationResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "isDemo": {
      "default": false,
      "description": "True for clearly labelled development fixtures",
      "title": "Isdemo",
      "type": "boolean"
     },
     "language": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Language"
     },
     "repoName": {
      "title": "Reponame",
      "type": "string"
     },
     "repoUrl": {
      "title": "Repourl",
      "type": "string"
     },
     "status": {
      "description": "awaiting_analysis | analyzing | analyzed | failed",
      "title": "Status",
      "type": "string"
     },
     "submittedAt": {
      "format": "date-time",
      "title": "Submittedat",
      "type": "string"
     },
     "targetRole": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Targetrole"
     }
    },
    "required": [
     "id",
     "candidateName",
     "candidateGithubHandle",
     "repoUrl",
     "repoName",
     "status",
     "submittedAt"
    ],
    "title": "CandidateRepoDetailResponse",
    "type": "object"
   },
   "CodeEvaluationResponse": {
    "description": "A written/scored evaluation. Scores are optional model/analyst opinions, never verdicts.",
    "properties": {
     "architectureScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Architecturescore"
     },
     "compositeHealthScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Compositehealthscore"
     },
     "concurrencySafetyScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Concurrencysafetyscore"
     },
     "evaluatedAt": {
      "format": "date-time",
      "title": "Evaluatedat",
      "type": "string"
     },
     "flaggedAnomalies": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Flaggedanomalies",
      "type": "array"
     },
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "maintainabilityScore": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Maintainabilityscore"
     },
     "recommendedDefenseTopics": {
      "items": {
       "type": "string"
      },
      "title": "Recommendeddefensetopics",
      "type": "array"
     },
     "repoId": {
      "format": "uuid",
      "title": "Repoid",
      "type": "string"
     },
     "rubricBreakdown": {
      "additionalProperties": true,
      "title": "Rubricbreakdown",
      "type": "object"
     },
     "summaryVerdict": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Summaryverdict"
     }
    },
    "required": [
     "id",
     "repoId",
     "evaluatedAt"
    ],
    "title": "CodeEvaluationResponse",
    "type": "object"
   },
   "ErrorResponse": {
    "description": "Every non-2xx response has this shape.",
    "properties": {
     "code": {
      "title": "Code",
      "type": "string"
     },
     "detail": {
      "title": "Detail",
      "type": "string"
     },
     "errors": {
      "anyOf": [
       {
        "items": {
         "$ref": "#/components/schemas/FieldError"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Errors"
     }
    },
    "required": [
     "detail",
     "code"
    ],
    "title": "ErrorResponse",
    "type": "object"
   },
   "FieldError": {
    "properties": {
     "field": {
      "title": "Field",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     }
    },
    "required": [
     "field",
     "message"
    ],
    "title": "FieldError",
    "type": "object"
   },
...
   "RepoSubmissionRequest": {
    "properties": {
     "branch": {
      "anyOf": [
       {
        "maxLength": 100,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "Null = repository default branch",
      "title": "Branch"
     },
     "candidateAvatarUrl": {
      "anyOf": [
       {
        "maxLength": 512,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Candidateavatarurl"
     },
     "candidateGithubHandle": {
      "maxLength": 120,
      "minLength": 1,
      "title": "Candidategithubhandle",
      "type": "string"
     },
     "candidateName": {
      "maxLength": 255,
      "minLength": 1,
      "title": "Candidatename",
      "type": "string"
     },
     "repoUrl": {
      "description": "https://github.com/<owner>/<repo>",
      "maxLength": 512,
      "title": "Repourl",
      "type": "string"
     },
     "targetRole": {
      "anyOf": [
       {
        "maxLength": 150,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Targetrole"
     }
    },
    "required": [
     "candidateName",
     "candidateGithubHandle",
     "repoUrl"
    ],
    "title": "RepoSubmissionRequest",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Synapse.AI",
  "version": "3.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/assessment/repo/recent": {
   "get": {
    "operationId": "get_recent_repos_api_v1_assessment_repo_recent_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 20,
       "maximum": 50,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/CandidateRepoDetailResponse"
         },
         "title": "Response Get Recent Repos Api V1 Assessment Repo Recent Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Submitted repositories with their measured results, if any (newest first)",
    "tags": [
     "Code Verifier"
    ]
   }
  },
  "/api/v1/assessment/repo/submit": {
   "post": {
    "operationId": "submit_repo_api_v1_assessment_repo_submit_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RepoSubmissionRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CandidateRepoDetailResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "Register a repository for analysis (stored as awaiting_analysis; nothing is scored)",
    "tags": [
     "Code Verifier"
    ]
   }
  },
  "/api/v1/assessment/repo/{repo_id}": {
   "get": {
    "operationId": "get_repo_detail_api_v1_assessment_repo__repo_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "repo_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Repo Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CandidateRepoDetailResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "404": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Not found"
     },
     "409": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Conflict"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Validation failed"
     },
     "500": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "Unexpected server error"
     },
     "503": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorResponse"
        }
       }
      },
      "description": "A dependency (AI provider or database) is unavailable"
     }
    },
    "summary": "One repository and its results",
    "tags": [
     "Code Verifier"
    ]
   }
  },
...
 }
}